Slightly restructured code, added Scenes
