+++ MeshAnimation/MeshAnimation.csproj
@@ -91,7 +91,9 @@
     <Compile Include="Clustering\KMeansUtils.cs" />
     <Compile Include="Config.cs" />
     <Compile Include="DataStructures\KDTree.cs" />
-    <Compile Include="MathUtil\BasicProcessing.cs" />
+    <Compile Include="Scenes\DemoScene.cs" />
+    <Compile Include="Scenes\Scene.cs" />
+    <Compile Include="Util\MeshProcessing.cs" />
     <Compile Include="Mode.cs" />
     <Compile Include="Optimization\IOptimizer.cs" />
     <Compile Include="Optimization\SSDROptimizer.cs" />
